Moved intersection calculation to a common place to be used in the update process
