+++ dsas-wps/src/main/java/gov/usgs/cida/dsas/wps/NearestPointOnLineProcess.java
@@ -5,6 +5,7 @@ import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.geom.LineSegment;
 import com.vividsolutions.jts.geom.LineString;
 import com.vividsolutions.jts.geom.MultiLineString;
+import gov.usgs.cida.dsas.wps.geom.IntersectionCalculator;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.geoserver.wps.gs.GeoServerProcess;
@@ -84,7 +85,7 @@ public class NearestPointOnLineProcess implements GeoServerProcess {
                         int gCount = geometry.getNumGeometries();
                         for (int gIndex = 0; gIndex < gCount; ++gIndex) {
                             LineString string = (LineString)geometry.getGeometryN(gIndex);
-                            for (LineSegment currentSegment : CreateTransectsAndIntersectionsProcess.toLineSegments(string)) {
+                            for (LineSegment currentSegment : IntersectionCalculator.toLineSegments(string)) {
                                 double currentDistance = currentSegment.distance(intpuCoordinateL);
                                 if (currentDistance < closestDistance) {
                                     closestDistance = currentDistance;
